requirements.txt
numpy>=1.16
torch==1.5.0
fair-esm==0.3.1
pandas>=1.2.3
tqdm>=4.60.0
transformers==4.5.1